Update and rename nic/update.php to update.php
too many changes to list, but was mainly changed by rearranging the request validation to verify and stop processing invalid requests before touching the cPanel server.

removed the changes that were made to eventually manage IPv6 only because that would need to be part of a bigger picture for other protocols since dyndns2 protocol doesn't support IPv6 as far as I know. We can add it back in if there is additional protocol support added back in.

The queryApi helper was modified to take an authString instead which is formed inside an if block early in the code. This was done to support Basic or cpanel authentication.